requirements.txt
alpaca-py>=0.21.0
yfinance>=0.2.36
gnews>=0.3.7
feedparser>=6.0.11
vaderSentiment>=3.3.2
pandas-ta>=0.3.14b
pandas>=2.0.0
apscheduler>=3.10.4
streamlit>=1.32.0
loguru>=0.7.2
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
plotly>=5.18.0
